requirements.txt
numpy
matplotlib
seaborn
pims
nd2reader
tifffile
scikit-image
cellpose